+++ internal/proposal/service.go
@@ -7,7 +7,10 @@ import (
 	"github.com/google/uuid"
 	coresdk "github.com/goverland-labs/goverland-core-sdk-go"
 	coreproposal "github.com/goverland-labs/goverland-core-sdk-go/proposal"
+	"github.com/goverland-labs/inbox-api/protobuf/inboxapi"
+	"google.golang.org/grpc"
 
+	"github.com/goverland-labs/inbox-web-api/internal/auth"
 	"github.com/goverland-labs/inbox-web-api/internal/entities/dao"
 	"github.com/goverland-labs/inbox-web-api/internal/entities/proposal"
 )
@@ -15,6 +18,10 @@ import (
 type DataProvider interface {
 	GetProposal(ctx context.Context, id string) (*coreproposal.Proposal, error)
 	GetProposalList(ctx context.Context, params coresdk.GetProposalListRequest) (*coreproposal.List, error)
+}
+
+type AIProvider interface {
+	GetAISummary(ctx context.Context, in *inboxapi.GetAISummaryRequest, opts ...grpc.CallOption) (*inboxapi.GetAISummaryResponse, error)
 }
 
 type DaoProvider interface {
@@ -25,13 +32,15 @@ type Service struct {
 	cache *Cache
 	dp    DataProvider
 	dao   DaoProvider
+	aip   AIProvider
 }
 
-func NewService(cache *Cache, dp DataProvider, dao DaoProvider) *Service {
+func NewService(cache *Cache, dp DataProvider, dao DaoProvider, aip AIProvider) *Service {
 	return &Service{
 		cache: cache,
 		dp:    dp,
 		dao:   dao,
+		aip:   aip,
 	}
 }
 
@@ -89,4 +98,17 @@ func (s *Service) GetList(ctx context.Context, ids ...string) ([]*proposal.Propo
 	}
 
 	return hits, nil
+}
+
+// GetAISummary request AI summary from storage and wrap to MD format
+func (s *Service) GetAISummary(ctx context.Context, sess auth.Session, proposalID string) (string, error) {
+	summary, err := s.aip.GetAISummary(ctx, &inboxapi.GetAISummaryRequest{
+		UserId:     sess.UserID.String(),
+		ProposalId: proposalID,
+	})
+	if err != nil {
+		return "", fmt.Errorf("get ai summary: %w", err)
+	}
+
+	return fmt.Sprintf("# AI summary\n\n%s", summary.GetSummary()), nil
 }